added requieredAxisLabelModulus to ChartXAxis
the requieredAxisLabelModulus makes sure that label at certain indexes
are always drawn if possible. E. g. if the chart wants to show a few
days with hours, but always wants to show the start of a day. In this
case requieredAxisLabelModulus would be set to 24 so that at least
every 24th label is drawn.