Enhancement - Added the parameter eventParentEnabled -  🤓 👻
In my application I have to use the horizontal scroll to show a menu on a view which has a pie o radar chart, I can disable the gestures on the rest of the charts but in the case of Pie or Radar, I can't. o I added the parameter just to know if the user want to the chart rotate or use the scroll for anything else.