Add new TimeLine chart type
This adds a new chart type in the demo app to showcase a line chart with values
on the x-axis represented as Doubles not indices/labels.

- Code sign identity and customizing controller

Set standard generalized code sign identity values

- Added data pipeline for extra numeric data for the X axis

This change adds a pipeline to get the x values as Doubles from the
ViewController to the appropriate renderer.  This extra data will
allow more precise placement of points in graphs.  However, this extra
data isn’t used for rendering yet.

- Put support for xNumericVal to the ChartData classes

The extra data for the x-axis is called xNumericVal.  It lives in ChartDataEntry
to make it available to all chart types.  It makes sense that Base
Chart Data classes would support xNumericVal as well.

The ViewController currently uses _deltaX (to determine max visible range of
x values, etc).  The ViewController needs to access xNumericValMin/Max for
to properly calculate deltaX.

- Add util for testing equality for Doubles

With the addition of Double values on the xAxis testing for equality of Doubles
will be used more.  Making this functionality a part of the ChartUtils framework.