+ aggregated bar chart
The main idea of the chart is to add possibility visually aggregate
rendered data on the chart, e.g. on such data like temperature changes
during some period of time it is usefull visually agregate the values
under one bar and show extrime values (min/max or median), for now, it
is just max value of data entries which were aggreageted. And
expand/disclose the data visually on the scale of timeline.