Added ability to disable/enable parent scroll view on drag
The default behavior was to disable parent scroll view on drag to avoid touch conflicts between chart view and  parent scroll view.

However if your dragging is horizontal and scrolling is vertical only (and vice versa with horizontal bar chart view), it’s desirable to disable this behavior to allow scroll view to scroll in the opposite direction.