Updated LineChart to Support an Array of Colors for CircleHoles
Updated LineChartDataSet Interface to contain an array of colors for circleHoles and a function to get a circleHole color at a given index.
Updated LineChartDataSet Implementation file to instantiate an empty NSUIColor array for circleHoleColors.
Added implementation to get a color out of the new circleHoleColors array by index, returning the existing circleHoleColor property if the array is empty or the index is out of bounds.
Updated the call for drawing a circleHole in the LineChartRenderer to use the new function to get the circleHoleColor by index instead of referencing the property