Fix for issue where data points are not highlighted
The `entriesForXValue` func finds the data entries for a given x-axis
value. It used an optimization where it would do a binary search to find
the first matching entry then take a slice of data stopping when no more
entries matched.

The optimization broke the highlight feature because data points that
should have matched were not matching resulting in no entries being
returned.

This commit replaces the optimized code with a call to `.filter` which
will be slower for larger data sets but correctly returns filtered
entries.