fix(ChartDataSet): highlight not being invoked for expected user interactions(FIxes #4719)
Changed the entriesForXValue and entryIndex method in the ChartDataSet
class to resolve #4719. A few things were happening, entriesForXValue
was using a predicate for the partition function that would never find
the value it was looking for as it is giving you the correct location in
the tree to start your search and not actually searching; hence why the
values are rarely selected. The prefix array method for lazy rendering
views also needs a different match predicate since the partitioned array
has values that wont need to be highlighted. There is another additional
change I made in this PR related to the closest logic for entryIndex;
prior to this change it was grabbing the next index and not the closest.
Example of the incorrect closest value: xValue = 6.2 would return back
closest as 7 or even 12 whichever one was directly preceding its own
index, now it returns the shortest distance between x values so it can
be more accurate to the users touch.